Manpower Changed As per Table

+++ src/Modules/Execution/Execution.Application/Features/Manpower/Commands/ManpowerCommands.cs
@@ -4,5 +4,5 @@ using Himapp.Execution.Application.Features.Manpower.Models;
 namespace Himapp.Execution.Application.Features.Manpower.Commands;
 
 public sealed record CreateManpowerCommand(CreateManpowerRequest Request) : IRequest<ManpowerModel>;
-public sealed record UpdateManpowerCommand(long Id, UpdateManpowerRequest Request) : IRequest<ManpowerModel?>;
-public sealed record DeleteManpowerCommand(long Id) : IRequest<bool>;
+public sealed record UpdateManpowerCommand(int Id, UpdateManpowerRequest Request) : IRequest<ManpowerModel?>;
+public sealed record DeleteManpowerCommand(int Id) : IRequest<bool>;
